+++ NEWS.md
@@ -1,5 +1,6 @@
 # shinymanager 1.0.110
 
+* Can configure shiny input for editing user in `secure_app()` with new `inputs_list`.
 * `auth_ui()` now accept a `background` argument change default css.
 
 # shinymanager 1.0.100